+++ flow-typed/react-virtualized.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+declare module 'react-virtualized/dist/commonjs/List' {
+  declare export default class List extends React.PureComponent {}
+}