+++ packages/interweave-emoji-picker/src/SearchBar.js
@@ -52,9 +52,9 @@ export default class SearchBar extends React.PureComponent<SearchBarProps, Searc
    * When the parent `Picker` search query is reset, also reset the input field.
    */
   componentWillReceiveProps({ searchQuery }: SearchBarProps) {
-    if (searchQuery === '') {
+    if (searchQuery === '' && this.props.searchQuery) {
       this.setState({
-        query: '',
+        query: searchQuery,
       });
     }
   }
@@ -102,7 +102,7 @@ export default class SearchBar extends React.PureComponent<SearchBarProps, Searc
           className={classNames.searchInput}
           placeholder={messages.search}
           ref={this.handleRef}
-          type="search"
+          type="text"
           value={this.state.query}
           onChange={this.handleChange}
           onKeyUp={onKeyUp}