+++ packages/interweave-emoji-picker/tests/mocks.js
@@ -1,0 +1,7 @@
+import { CONTEXT_CLASSNAMES, CONTEXT_MESSAGES } from '../src/constants';
+
+// eslint-disable-next-line
+export const PICKER_CONTEXT = {
+  classNames: CONTEXT_CLASSNAMES,
+  messages: CONTEXT_MESSAGES,
+};